+++ changelog.md
@@ -140,6 +140,7 @@
 - `--hint:processing` is now supported and means `--hint:processing:on`
   (likewise with other hints and warnings), which is consistent with all other bool flags.
   (since 1.3.3).
+- `nim doc -r main` and `nim rst2html -r main` now call openDefaultBrowser
 
 ## Tool changes